+++ CHANGELOG.md
@@ -12,25 +12,40 @@ All notable changes to this project will be documented in this file. For commit 
    - The `Q` shortcut in audio files will toggle the new panel visibility.
    - `E` shortcut to change between tabs of the panel if it's open.
  - Api log filtering support via `apiFilter` option for logging.
-   - regex filter that excludes matching full api paths from being logged. (eg. '/user\?id\=self') Defaults to '^/health|^/favicon.ico|^/static|^/public/static'
+   - regex filter that excludes matching full api paths from being logged. (eg. `/user\?id\=self`) Defaults to `^/health|^/favicon.ico|^/static|^/public/static`
    - Add config option to disable /health endpoint logging (#2291)
  - More shortcuts (#2300)
    - `CTRL+F1` switch to normal view.
    - `CTRL+F2` switch to gallery view.
    - `CTRL+F3` switch to list view.
    - `Double click` on empty space to select all items.
- - Wildcard search support (limited to sqlite GLOB)
+ - Option to hide certain files in UI by extension (#2403) (#2277)
+ - Aditional search features
+   - wildcard search similar to regex, but limited to sqlite GLOB syntax (#769)
+   - advanced search tool that shows all results in a listing thats sortable and actionable like a normal listing (#1051)
+   - advanced search supports multiple custom scopes per search
+   - multiple search terms with AND or OR logic support
 
  **Notes**:
  - Enhanced indexing scheduler which doesn't wake the disk as often.
  - New API route `media/lyrics` used to fetch and parse lyrics (embedded or from `.lrc` sidecar) (#2360)
  - Swiping down gesture in fullscreen videos exit fullscreen instead of close preview.
  - Improved styles for path selection and tables
+ - Edit Sidebar links has new "show tools in sidebar" toggle and all users have this enabled by default. can be disabled via for new users `userDefaults.showToolsInSidebar: false`
 
  **BugFixes**:
  - Blue overlay when using gestures in video files on mobile (#2360)
  - Playback queue wasn't updating when changing of folder (#2360)
  - Added missing `UserDefaults.ShowCopyPath` option to config file (#2364)
+
+## v1.3.10
+
+ **Notes**:
+ - When indexing is disabled for a source, the usage will always reported as partition size.
+ - removed deprecated `source.config.disableIndexing`, see [rules](https://filebrowserquantum.com/en/docs/advanced/source-configuration/conditional-rules/#disable-indexing)
+
+ **BugFixes**:
+ - Disable index option not working (#2385)
 
 ## v1.3.9